Drug Calculated Properties

Drug properties that have been predicted by ChemAxon or ALOGPS based on the imputed chemical structure. Associated links below will redirect to descriptions of the specific term.

Calculated_Properties_Drug

Format

a tibble with 3 variables:

kind

Name of the property

value

Predicted physicochemical properties; obtained by the use of prediction software such as ALGOPS and ChemAxon

source

Name of the software used to calculate this property, either ChemAxon or ALOGPS

parent_key

drugbank id

Source

Drugbank Documentation

Details

Each drug may have one or more calculated property.

The following calculated properties are provided:

Property	Description

logP	The predicted partition coefficient (LogP) based on the ratio of solubility of the molecule in 1-octanol compared to water; predicted by ALOGPS

logS	The predicted solubility (LogS) of the molecule; predicted by ALOGPS.

IUPAC Name	The predicted International Union of Pure and Applied Chemistry (IUPAC) nomenclature for the structure; predicted by ChemAxon

Traditional IUPAC Name	The non-systematic (or common) name for the molecule, which is not recognized by any formal nomenclature system; imported from ChemAxon.

Molecular Weight	The predicted ratio of the average mass of one molecule of an element or compound to one twelfth of the mass of an atom of carbon-12; calculated by ChemAxon.

Monoisotopic Weight	The predicted mass of the most abundant isotope of the drug; calculated by ChemAxon.

SMILES	The simplified molecular-input line-entry system (SMILES) is a line notation used for describing the structure of chemical species using short ASCII strings; calculated by ChemAxon.

InChI	A prediction of the IUPAC International Chemical Identifier (InChI); imported by ChemAxon.

InChIKey	The condensed digital representation of the IUPAC International Chemical Identifier (InChI); imported by ChemAxon.

Polar Surface Area (PSA)	A descriptor, based on the polarized atoms of the molecule, that allows estimation of transport properties and of the passive molecular transport through membranes of the drug; predicted by ChemAxon.

Refractivity	The predicted molar refractivity of the molecule, which is strongly related to the volume of the molecules and to London dispersive forces that play crucial part in drug-receptor interactions; predicted by ChemAxon.

Polarizability	The predicted relative tendency of the electron cloud (charge distribution) of the molecule to be distorted by an external electric field; polarizability values predicted by ChemAxon.

Rotatable Bond Count	The predicted number of rotatable bonds in the molecule; predicted by ChemAxon. Unsaturated bonds, and single bonds connected to hydrogens or terminal atoms, single bonds of amides, sulphonamides and those connecting two hindered aromatic rings (having at least three ortho substituents) are considered non-rotatable.

H Bond Acceptor Count	A calculation of the sum of the hydrogen bond acceptor atoms. An acceptor atom always has a lone electron pair/lone electron pairs that is capable of establishing a H bond. Predicted by ChemAxon.

H Bond Donor Count	A calculation of the sum of the atoms in the molecule which have hydrogen bond donor property. Predicted by ChemAxon.

pKa (strongest acidic)	The strongest acidic pka value of the molecule; predicted by ChemAxon.

pKa (strongest basic)	The strongest basic pka value of the molecule; predicted by ChemAxon.

Physiological Charge	Charge of the molecule at physiological pH; predicted by ChemAxon.

Number of Rings	A calculation of the number of rings in the molecule; predicted by ChemAxon.

Bioavailability	Fraction of administered dose that is predicted to reach the systemic circulation; predicted by ChemAxon.

Rule of Five	A reflection of the absorption or permeation of a molecule; considered “yes” when the molecular weight is under 500 g/mol, the value of logP is lower than 5, and the molecule has utmost 5 H-donor and 10 H-acceptor atoms; predicted by ChemAxon.

Ghose Filter	A filter that defines drug-likeness constraints as follows: calculated log P is between -0.4 and 5.6, molecular weight is between 160 and 480, molar refractivity is between 40 and 130, and the total number of atoms is between 20 and 70. Imported from ChemAxon.

MDDR-Like Rule	Indicates compliance of drug-like characteristics based on number of rings, rigid bonds and rotatable bonds; calculated by ChemAxon.