Carriers/ Enzymes/ Targets/ Transporters

Protein targets of drug action, enzymes that are inhibited/induced or involved in metabolism, and carrier or transporter proteins involved in movement of the drug across biological membranes.

Carriers_Drug

Enzymes_Drug

Targets_Drug

Transporters_Drug

Format

a tibble with 6 variables:

id

Identifier for the record

name

related name

organism

Organism that the protein comes from.

known_action

Whether the pharmacological action of the drug is due to this target interaction.

position

related position

parent_id

drugbank id

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 805 rows and 6 columns.

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 5158 rows and 8 columns.

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 18780 rows and 6 columns.

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 3050 rows and 6 columns.

Source

Drugbank Documentation

Details

Each of targets, enzymes, carriers and transporters contain one or more child elements tibbles