External Identifiers for Drugs/ Carriers/ Enzymes/ Targets/ Transporters

Identifiers used in other websites or databases providing information about this drug.

External_Identifiers_Drug

External_Identifiers_Polypeptide_Carrier_Drug

External_Identifiers_Polypeptide_Enzyme_Drug

External_Identifiers_Polypeptide_Target_Drug

External_Identifiers_Transporter_Drug

Format

a tibble with following features:

resource

Name of the source database

identifier

Identifier for this drug in the given resource

parent_key

drugbank key

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 82974 rows and 3 columns.

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 424 rows and 3 columns.

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 2365 rows and 3 columns.

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 21770 rows and 3 columns.

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 1287 rows and 3 columns.

Source

Drugbank Documentation

Details

Each drug may have one or more external identifier.

Drug identifiers may be provided for the following resources

Wikipedia

ChEBI

ChEMBL

PubChem Compound

PubChem Substance

Drugs Product Database (DPD)

KEGG Compound

KEGG Compound

KEGG Drug

ChemSpider

BindingDB

National Drug Code Directory

GenBank

PharmGKB

PDB

Guide to Pharmacology

ZINC

RxCUI